refactor(file-utils): move file util functions to own module
